feat: move dashboard metrics endpoint to dashboboard plugin
